refactor: split FAQ category tasks
